huge classes decomposed #SCL-15807
